Removes views from the General grouping
